+++ settlement-management-sheet/client/src/app/thunks/glossary/subtypes/properties/removeSubTypePropertyThunk.ts
@@ -11,9 +11,11 @@ import {
   SubTypeProperty,
 } from '@/app/slice/subTypeSlice.js';
 import deleteSubTypePropertyService from '@/services/glossary/subTypes/deleteSubTypePropertyService.js';
-import { deleteSubTypeProperty } from '@/app/commands/subtype.ts';
+import { deleteSubTypeProperty } from '@/app/commands/userSubtype.ts';
 import getSubTypeStateAtKey from '@/utility/dataTransformation/getSubTypeStateAtKey.ts';
 import checkAdmin from '@/utility/security/checkAdmin.ts';
+import { logger } from '@/utility/logging/logger.ts';
+import { sysDeleteSubTypeProperty } from '@/app/commands/sysSubtype.ts';
 
 export function deleteSubTypePropertyThunkRoot({
   propertyId,
@@ -26,12 +28,8 @@ export function deleteSubTypePropertyThunkRoot({
       propertyId
     );
     if (!property) {
-      dispatch(
-        showSnackbar({
-          message: 'How can I delete nothing? Property not found.',
-          type: 'error',
-          duration: 3000,
-        })
+      logger.snError(
+        'If I told you it was already gone, would you believe me? Please report this bug.'
       );
       return;
     }
@@ -45,7 +43,11 @@ export function deleteSubTypePropertyThunkRoot({
         })
       );
 
-      await deleteSubTypeProperty({ id: propertyId });
+      if (property.system) {
+        await sysDeleteSubTypeProperty({ id: propertyId });
+      } else {
+        await deleteSubTypeProperty({ id: propertyId });
+      }
 
       dispatch(
         addDirtyKeypath({
@@ -58,12 +60,9 @@ export function deleteSubTypePropertyThunkRoot({
       dispatch(
         addSubTypeProperty({ properties: [property], system: property.system })
       );
-      dispatch(
-        showSnackbar({
-          message: 'Error removing property. Try again later.',
-          type: 'error',
-          duration: 3000,
-        })
+      logger.snError(
+        'The property wanted to say something before being obliterated. Or it bribed the server. Please report.',
+        { error }
       );
     }
   };
